backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.1

# AWS
boto3==1.29.7
django-storages==1.14.2

# API & Validation
drf-yasg==1.21.7
django-filter==23.5
Pillow==10.1.0

# Caching
django-redis==5.4.0
redis==5.0.1

# Task Queue (for async tasks)
celery==5.3.4
django-celery-beat==2.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0

# Monitoring
sentry-sdk==1.38.0

# AI/ML (for future health insights)
numpy==1.26.2
pandas==2.1.3

